# front-openapi/src/api/generated/models/ProductMiniResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductMiniResponse = {
    id?: number;
    name?: string;
    price?: number;
};