# front-openapi/src/api/generated/models/ProductResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProductResponse = {
    id?: number;
    name?: string;
    description?: string;
    category?: string;
    brand?: string;
    price?: number;
    quantity?: number;
    image?: string;
};